+++ service/controller/resource/clustersecret/desired.go
@@ -98,6 +98,12 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) ([]*cor
 
 		noProxy := noProxy(cr, r.proxy.NoProxy)
 
+		// The three values below are specific to cert-manager because
+		// we use upstream chart schema.
+		values["no_proxy"] = noProxy
+		values["http_proxy"] = r.proxy.HttpProxy
+		values["https_proxy"] = r.proxy.HttpsProxy
+
 		values["cluster"] = map[string]interface{}{
 			"proxy": map[string]string{
 				"noProxy": noProxy,